third_party/blink/web_tests/wpt_internal/css/css-contain/parsing/contain-computed.html

<!doctype html>
<meta charset="utf-8">
<title>CSS Containment Module: getComputedStyle().contain</title>
<!-- Merge these tests into upstream wpt test if/when added to the css-contain spec -->
<link rel="help" href="https://drafts.csswg.org/css-contain/#contain-property">
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/1031">
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/5796">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<div id="target"></div>
<script>
test_computed_value("contain", "block-size");
test_computed_value("contain", "inline-size");
</script>
